requirements.txt
#嘻嘻
transformers==4.30.2
tts==0.20.1
soundfile==0.12.1
torchaudio==2.0.2
numpy==1.24.3